ділення цілих чисел - втрата точності

+++ homework_05/src/telemetry.cpp
@@ -87,7 +87,7 @@ if (field_count != EXPECTED_FIELD_COUNT) {
 double compute_frame_rate_hz(const Frame frames[], int frame_count) {
     const long elapsed_ms = frames[frame_count - 1].timestamp_ms - frames[0].timestamp_ms;
 
-    return static_cast<double>((frame_count - 1) * 1000 / elapsed_ms);
+return static_cast<double>(frame_count - 1) * 1000.0 / static_cast<double>(elapsed_ms);
 }
 
 int read_frames(const char* path, Frame frames[], int max_frames) {
